+++ extensions/moneytree/CHANGELOG.md
@@ -2,5 +2,7 @@
 
 ## [Initial Version] - {PR_MERGE_DATE}
 
-- Added authentication using OAuth 2.0 with PKCE
+- Added authentication using email and password + OAuth 2.0 (PKCE) under the hood
 - Added commands to list credentials, accounts with balances and latest transactions
+- Added actions to copy details and logout to all commands
+- Added Windows support for logout shortcut